[Chore] Cleanup
- Created new package `auto_update`. Moved related methods and structs to that folder. This makes main a bit more clean as these methods don't deal with the fundamental aspect of the go-work-tracker
- Changed `go.mod` module name from `go-work-tracker` to `github.com/theBGuy/go-work-tracker` this seems to be the standard way to do it
